Refactored.

+++ smc/smc.h
@@ -99,11 +99,11 @@ typedef struct {
     int number;
     UInt8 id;
     SMCBytes_t name;
-    float actual_speed;
-    float minimal_speed;
-    float maximum_speed;
-    float safe_speed;
-    float target_speed;
+    int actual_speed;
+    int minimal_speed;
+    int maximum_speed;
+    int safe_speed;
+    int target_speed;
     int mode;
 } Fan_t;
